backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg[binary]==3.2.3
python-multipart==0.0.12
httpx==0.27.2
paho-mqtt==2.1.0
